requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.2

# Visualization
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# Preprocessing & Feature Engineering
scikit-learn==1.5.0
imbalanced-learn==0.12.3

# Association Rules (Apriori / FP-Growth)
mlxtend==0.23.1

# Clustering
scipy==1.13.1

# Time Series Forecasting
statsmodels==0.14.2
prophet==1.1.5

# Notebook automation
papermill==2.6.0
jupyterlab==4.2.2
ipykernel==6.29.4

# Reporting
jinja2==3.1.4
